Added no AP firmware...
Normally when the connection to the main wifi AP is lost, the software goes into AP mode so it can be reconfigured if needed. This is turned off, and will keep trying to reconnect to the wifi AP that it is configured for.  This is done because with some configurations going into AP mode crashes the ESP32, and it will not come back up unless manually reset.

For this version the only way to change the AP is to ground GPIO 3 for 5 seconds to put the ESP32 into AP mode.